moved filters folder into shared/feature
